# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Simar's personal dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Quote {
  id: number;
  text: string;
  category: string;
  /** @nullable */
  author?: string | null;
}

export interface CategoryProgress {
  category: string;
  label: string;
  emoji: string;
  percentage: number;
  completedItems: number;
  totalItems: number;
}

export interface DashboardSummary {
  greeting: string;
  quote: Quote;
  progressRings: CategoryProgress[];
  streak: number;
  currentSemester: string;
  winsThisMonth: number;
  /** @nullable */
  dailyFocus: string | null;
}

export interface RoadmapItem {
  id: number;
  category: string;
  title: string;
  completed: boolean;
  order: number;
  createdAt: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  linkUrl?: string | null;
}

export interface RoadmapItemInput {
  category: string;
  title: string;
  order?: number;
}

export interface RoadmapItemUpdate {
  title?: string;
  completed?: boolean;
  order?: number;
  notes?: string;
  status?: string;
  contactEmail?: string;
  linkUrl?: string;
}

export interface Win {
  id: number;
  description: string;
  createdAt: string;
}

export interface WinInput {
  description: string;
}

export interface WinsSummary {
  wins: Win[];
  totalThisMonth: number;
  total: number;
}

export interface Milestone {
  id: number;
  title: string;
  /** Type: checkbox, progress, countdown, counter */
  milestoneType: string;
  completed: boolean;
  /** @nullable */
  targetCount?: number | null;
  /** @nullable */
  currentCount?: number | null;
  /** @nullable */
  targetDate?: string | null;
  createdAt?: string;
}

export interface MilestoneInput {
  title: string;
  milestoneType: string;
  targetCount?: number;
  targetDate?: string;
}

export interface MilestoneUpdate {
  title?: string;
  completed?: boolean;
  currentCount?: number;
  targetCount?: number;
  targetDate?: string;
}

export interface DailyFocus {
  id: number;
  text: string;
  date: string;
}

export interface DailyFocusInput {
  text: string;
}

export interface PlannerNote {
  id: number;
  dayOfWeek: string;
  content: string;
  weekStart: string;
}

export interface PlannerNoteInput {
  dayOfWeek: string;
  content: string;
  weekStart: string;
}

export interface PlannerNoteUpdate {
  content?: string;
}

export interface CalendarTask {
  id: number;
  title: string;
  date: string;
  /** @nullable */
  notes?: string | null;
  completed: boolean;
  createdAt: string;
}

export interface CalendarTaskInput {
  title: string;
  date: string;
  notes?: string;
}

export interface CalendarTaskUpdate {
  title?: string;
  date?: string;
  notes?: string;
  completed?: boolean;
}

export interface VisionItem {
  id: number;
  /** image or quote */
  type: string;
  /** URL for images, text for quotes */
  content: string;
  /** @nullable */
  caption?: string | null;
  createdAt: string;
}

export interface VisionItemInput {
  type: string;
  content: string;
  caption?: string;
}

export interface Document {
  id: number;
  category: string;
  name: string;
  lastUpdated: string;
  /** @nullable */
  notes?: string | null;
  /**
     * Object storage path for the uploaded file (e.g. /objects/uploads/uuid)
     * @nullable
     */
  fileUrl?: string | null;
}

export interface DocumentInput {
  category: string;
  name: string;
  notes?: string;
  /** Object storage path returned by the upload endpoint */
  fileUrl?: string;
}

export interface DocumentUpdate {
  name?: string;
  category?: string;
  notes?: string;
  lastUpdated?: string;
  /** Object storage path returned by the upload endpoint */
  fileUrl?: string;
}

export interface QuoteInput {
  text: string;
  category: string;
  author?: string;
}

export interface Settings {
  id: number;
  name: string;
  /** built-in wallpaper key or custom URL */
  wallpaper: string;
  blurAmount: number;
  darknessAmount: number;
  currentSemester: string;
}

export interface SettingsUpdate {
  name?: string;
  wallpaper?: string;
  blurAmount?: number;
  darknessAmount?: number;
  currentSemester?: string;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. /objects/uploads/uuid). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type GetRoadmapItemsParams = {
category?: string;
};

export type GetDocumentsParams = {
category?: string;
};

export type GetQuotesParams = {
category?: string;
};

export type GetRandomQuoteParams = {
category?: string;
};