# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Simar's personal dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /milestones:
    get:
      operationId: getMilestones
      tags: [milestones]
      summary: List all milestones
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a new milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      summary: Delete a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /planner:
    get:
      operationId: getPlannerNotes
      tags: [planner]
      summary: Get all weekly planner notes
      responses:
        "200":
          description: List of planner notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlannerNote"
    post:
      operationId: createPlannerNote
      tags: [planner]
      summary: Create a planner note for a day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerNote"
  /planner/{id}:
    patch:
      operationId: updatePlannerNote
      tags: [planner]
      summary: Update a planner note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerNoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerNote"
    delete:
      operationId: deletePlannerNote
      tags: [planner]
      summary: Delete a planner note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /calendar/tasks:
    get:
      operationId: getCalendarTasks
      tags: [calendar]
      summary: List dated calendar tasks
      responses:
        "200":
          description: Calendar tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarTask"
    post:
      operationId: createCalendarTask
      tags: [calendar]
      summary: Create a dated task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarTask"
  /calendar/tasks/{id}:
    patch:
      operationId: updateCalendarTask
      tags: [calendar]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarTask"
    delete:
      operationId: deleteCalendarTask
      tags: [calendar]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: List documents optionally filtered by category
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Add a document entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update document metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Milestone:
      type: object
      required: [id, title, milestoneType, completed]
      properties:
        id:
          type: number
        title:
          type: string
        milestoneType:
          type: string
          description: "Type: checkbox, progress, countdown, counter"
        completed:
          type: boolean
        targetCount:
          type: ["number", "null"]
        currentCount:
          type: ["number", "null"]
        targetDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    MilestoneInput:
      type: object
      required: [title, milestoneType]
      properties:
        title:
          type: string
        milestoneType:
          type: string
        targetCount:
          type: number
        targetDate:
          type: string
          format: date
    MilestoneUpdate:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean
        currentCount:
          type: number
        targetCount:
          type: number
        targetDate:
          type: string
          format: date
...
    PlannerNote:
      type: object
      required: [id, dayOfWeek, content, weekStart]
      properties:
        id:
          type: number
        dayOfWeek:
          type: string
        content:
          type: string
        weekStart:
          type: string
          format: date
    PlannerNoteInput:
      type: object
      required: [dayOfWeek, content, weekStart]
      properties:
        dayOfWeek:
          type: string
        content:
          type: string
        weekStart:
          type: string
          format: date
    PlannerNoteUpdate:
      type: object
      properties:
        content:
          type: string
    CalendarTask:
      type: object
      required: [id, title, date, completed, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        date:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CalendarTaskInput:
      type: object
      required: [title, date]
      properties:
        title:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string
    CalendarTaskUpdate:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string
        completed:
          type: boolean
...
    Document:
      type: object
      required: [id, category, name, lastUpdated]
      properties:
        id:
          type: number
        category:
          type: string
        name:
          type: string
        lastUpdated:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
          description: "Object storage path for the uploaded file (e.g. /objects/uploads/uuid)"
    DocumentInput:
      type: object
      required: [category, name]
      properties:
        category:
          type: string
        name:
          type: string
        notes:
          type: string
        fileUrl:
          type: string
          description: "Object storage path returned by the upload endpoint"
    DocumentUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        notes:
          type: string
        lastUpdated:
          type: string
          format: date
        fileUrl:
          type: string
          description: "Object storage path returned by the upload endpoint"
...