# lib/api-zod/src/generated/types/calendarTask.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Simar's personal dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarTask {
  id: number;
  title: string;
  date: Date;
  /** @nullable */
  notes?: string | null;
  completed: boolean;
  createdAt: Date;
}