# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Simar's personal dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  category: string;
  name: string;
  lastUpdated: Date;
  /** @nullable */
  notes?: string | null;
  /**
     * Object storage path for the uploaded file (e.g. /objects/uploads/uuid)
     * @nullable
     */
  fileUrl?: string | null;
}